Add Installer & Comments DOCBLOCK
Add Installer Controller, Installer layout and styler, and add comments
Docblock for PHPDOC

+++ gxadmin/inc/categories.php
@@ -1,6 +1,21 @@
+<?php
+/**
+* GeniXCMS - Content Management System
+* 
+* PHP Based Content Management System and Framework
+*
+* @package GeniXCMS
+* @since 0.0.1 build date 20150202
+* @version 0.0.1
+* @link https://github.com/semplon/GeniXCMS
+* @author Puguh Wijayanto (www.metalgenix.com)
+* @copyright 2014-2015 Puguh Wijayanto
+* @license http://www.opensource.org/licenses/mit-license.php MIT
+*
+*/?>
 <div class="row">
     <div class="col-md-12">
-        <h1>Categories 
+        <h1><i class="fa fa-cubes"></i>  Categories 
             <button class="btn btn-success pull-right" data-toggle="modal" data-target="#myModal">
                 <span class="glyphicon glyphicon-plus"></span> Add Category
             </button>
@@ -11,45 +26,49 @@
     <div class="col-sm-12">
         <div class="row">
             <?php
-                foreach ($data['cat'] as $c) {
-                    # code...
-                    // echo "<td>".$c->id."</td>";
-                    // echo "<td>".$c->name."</td>";
-                    // echo "<td>".$c->parent."</td>";
-                    // echo "<td></td>";
+                if($data['num'] > 0) {
+                    foreach ($data['cat'] as $c) {
+                        # code...
+                        // echo "<td>".$c->id."</td>";
+                        // echo "<td>".$c->name."</td>";
+                        // echo "<td>".$c->parent."</td>";
+                        // echo "<td></td>";
 
-                    if($c->parent == ""){
-                        echo "<div class=\"col-md-4 item\" >
-                        <div class=\"panel panel-default\">
-                            <div class=\"panel-heading\">
-                              <h3 class=\"panel-title\">{$c->name} 
-                              <a href=\"?page=categories&act=del&id={$c->id}\" class=\"pull-right\"
-                              onclick=\"return confirm('Are you sure you want to delete this item?');\">
-                              <span class=\"glyphicon glyphicon-remove\"></span></a></h3>
-                            </div>
-                            <div class=\"panel-body\">
-                            <ul class=\"list-group\">";
-                            foreach ($data['cat'] as $c2) {
-                                if($c2->parent == $c->id){
-                                    echo "<li class=\"list-group-item\">
-                                    <form action=\"\" method=\"POST\" name=\"updatecat\">
-                                    <div class=\"input-group\">
-                                        <a href=\"?page=categories&act=del&id={$c2->id}\" class=\"input-group-addon\"
-                                        onclick=\"return confirm('Are you sure you want to delete this item?');\"
-                                        ><span class=\"glyphicon glyphicon-remove\"></span></a>
-                                        <input type=\"text\" name=\"cat\" class=\"form-control\" value=\"{$c2->name}\">
-                                        <input type=\"hidden\" name=\"id\" value=\"{$c2->id}\">
-                                        <span class=\"input-group-btn\">
-                                            <button class=\"btn btn-default\" type=\"submit\" name=\"updatecat\">Go!</button>
-                                        </span>
-                                    </div>
-                                    </form>
-                                     </li>";
+                        if($c->parent == ""){
+                            echo "<div class=\"col-md-4 item\" >
+                            <div class=\"panel panel-default\">
+                                <div class=\"panel-heading\">
+                                  <h3 class=\"panel-title\">{$c->name} 
+                                  <a href=\"?page=categories&act=del&id={$c->id}\" class=\"pull-right\"
+                                  onclick=\"return confirm('Are you sure you want to delete this item?');\">
+                                  <span class=\"glyphicon glyphicon-remove\"></span></a></h3>
+                                </div>
+                                <div class=\"panel-body\">
+                                <ul class=\"list-group\">";
+                                foreach ($data['cat'] as $c2) {
+                                    if($c2->parent == $c->id){
+                                        echo "<li class=\"list-group-item\">
+                                        <form action=\"\" method=\"POST\" name=\"updatecat\">
+                                        <div class=\"input-group\">
+                                            <a href=\"?page=categories&act=del&id={$c2->id}\" class=\"input-group-addon\"
+                                            onclick=\"return confirm('Are you sure you want to delete this item?');\"
+                                            ><span class=\"glyphicon glyphicon-remove\"></span></a>
+                                            <input type=\"text\" name=\"cat\" class=\"form-control\" value=\"{$c2->name}\">
+                                            <input type=\"hidden\" name=\"id\" value=\"{$c2->id}\">
+                                            <span class=\"input-group-btn\">
+                                                <button class=\"btn btn-default\" type=\"submit\" name=\"updatecat\">Go!</button>
+                                            </span>
+                                        </div>
+                                        </form>
+                                         </li>";
+                                    }
                                 }
-                            }
-                        echo "</ul></div>";
-                        echo "</div></div>";
+                            echo "</ul></div>";
+                            echo "</div></div>";
+                        }
                     }
+                }else{
+                    echo "<div class=\"col-md-12\">No Categories Found</div>";
                 }
             ?>
         </div>
